add Task and TestCase models
These models will provide storage of tasks and test cases
for them. Both tables are linked by a one-to-many
relationship.

In the future, many-to-many communication is possible in
combination with algorithms to prevent duplicate test
data. This would not save performance, since the minimum
number of tests in simple tasks is repeated